chore(dockerfile) refactor dockerfile plugin into its own package
