+++ pkg/plugins/dockerfile/condition.go
