+++ pkg/plugins/dockerfile/test_fixtures/ARG.Dockerfile
