+++ pkg/plugins/dockerfile/test_fixtures/FROM.Dockerfile
